Move behaviours to dedicated module
