+++ mqtt-tester/src/main.rs
@@ -4,6 +4,7 @@
 //   file, You can obtain one at http://mozilla.org/MPL/2.0/.
 //
 
+mod behaviour;
 mod behaviour_test;
 mod client_report;
 mod command;
